refactor: abstract common command fields into Command.kt
